`comment-added` event can have multiple approvals
The `comment-added` event can include multiple approvals, but the
current implementation does not support this.  Refactor it to allow
multiple approvals to be handled.

Also add the description field in the approval object.

Refactor a bit to make the handling of supported approval categories
easier to extend later if necessary.