Fix event registration from other module
The purpose of the `register` decorator in the event factory is
to allow client scripts to define their own event classes and
have them dispatched when monitoring the event stream.

The current implementation does not work when an event class is
registered from another module, because it uses globals() which
only returns the symbols defined in the event module, so trying
to instantiate the object causes an exception.

Change the implementation to also store the module name when
registering, and then use that when instantiating the event
object.

Add unit tests to cover the event registration functionality.

Change-Id: I1af901621b07fc9236bc95a81956269507994871