More error handling improvements in stream
Handle IOError and ValueError separately from GerritError when
processing the steam.  In this case GerritError is raised if
there is a problem adding an event to the queue; it makes no
sense to then try to add an error event on the same queue.

Add separate handling of GerritError when running the gerrit
stream-events command.

Move the creation of the error event out to a utility method.

Change-Id: Iac5aa8db5ad1a444fd0e6217c32d1800f727276e