Refactor getting the Gerrit version
Move the functionality to get the Gerrit version from the main
client into the SSH client and reimplement it.

First attempt to get the version from the underlying Paramiko SSH
client.  If that fails, fall back to using the `gerrit version`
command as before.

Get and print the version in the example script.

Change-Id: Ia79853f80f898b43c301fcecee61fdfdb75e9811