Add a make target to build the source distribution
The `sdist` target invokes `python setup.py sdist` which builds the
source distribution .tar.gz file for the current release.

The build will fail if the package version specified in the module
is not the same as the tag at the current HEAD, or if the current
HEAD is not tagged with a properly formatted value.

Change-Id: Ib09f0b1a82f5ae49458f63d8fbeb9bb8bdd1cb2d