Don't put copyright and license headers in Python module docstrings
If the copyright and license headers are in module docstrings, they
appear in the generated API documentation pages and it looks crappy.

Instead, put the headers in regular Python comments.  Thus each file
still has the required copyright and license notice, but it is not
included everywhere in the generated documentation.

The documentation still includes an automatically generated copyright
notice at the bottom of each page.

Change-Id: Ibc5f4ca221c0a6de26f04f3b4511d8991e28b663