Add make target and rules to ensure environment setup tools are OK
To set up the environment properly, the host system needs to have the
virtualenv and pip packages already installed.  Add top level rules to
make sure the necessary executables exist.

When building the `sdist` target there is a check to make sure that the
git workspace does not have any untracked or modified files.  This can
cause the build to fail if the environment was set up with an old
version of virtualennv which does not include bundled versions of the
`distribute` and `setuptools` packages, and downloads them into the
working directory.

Add a new target `valid-virtualenv` which will fail the build if the
version of virtualenv is older than expected.  Add this target as a
dependency of the `sdist` target to force source distribution builds
to be done with an explicitly versioned virtualenv.

Change-Id: I1a78e8859512696005ab9d6dd790cfc5256dc2e3