Suppress pylint warning about catching too general exception
Commit 06c88c8 (merged with pull request #13) added exception
handling:

  except Exception as e:

which causes a pylint warning:

  W0703: Catching too general exception Exception

However in this case we actually do want to catch all types
of exception because the intention is to make sure pygerrit
doesn't fall over when something unexpected happens.

Add a suppression of this warning.  The suppression is only
effective on this particular line; the warning will be raised
again if we do the same thing again somewhere else.

Change-Id: If812d441c1dc625651db34c6d3763cedc98b68a9