Fix README markup to work properly on Gitlab
The markup used to show code snippets renders properly on GitHub
but on gitlab the snippets are not shown.

Change the markup so that it works on both.

With this change we lose the syntax highlighting on Github, but
this is a fair tradeoff if it means we can view it properly on
Gitlab.

Change-Id: Ie2faa0862fd467a806041b57aafd5b797e1a5cd2