Add back the ReStructured Text version of the README
It turns out that markdown is not supported in python packaging
and the README is not rendered properly on pypi.python.org.

Add back the RST version. Keep the MD version so we still have
something that looks OK in Github and Gitiles. The README is not
likely to be updated much so it shouldn't be too much overhead.

Change-Id: I68145a1fb3a6725c4e72bc60841286793ee91187