rest: handle endpoints that don't output json
The /changes/<id>/revisions/current/patch?zip endpoint doesn't return
json output but a binary data (the zipped contents of the change). The
usual get() function raises a ValueError when faced with this. As a
workaround provide a get_raw() function to allow callers to handle the
decoding of data.