Fixes #18: Explicitly set content type header when posting review
Since f37293a ("Fix a bug preventing file upload in PUT and POST"),
the Content-Type header is only set when the `data` argument is a
`dict`. However when called from the `review` method, it is a string.

Fix this by explicitly setting the Content-Type header in the call
to `review`.

Change-Id: I7346129776739a17d3a4ff1f5914f7ecf1af903d