Allow override of kwargs in get
The kwargs member contains the header
  'Accept': 'application/json'

This is not compatible with the content api of Gerrit. The content api
returns the content of a file as 'text/plain' encoded with base64. The
accept json header will confuse Gerrit to return the content of a file
as json.

This changes the get implementation to update the internal kwargs with
the kwargs from the function parameters. This makes it possible to
override the 'Accept' header (or any other parameter).

Example:

  response = rest.get(
      'projects/someproject/branches/master/files/foo.png/content',
      headers={'Accept': 'text/plain'})
  content = base64.b64decode(response)

Change-Id: I0d4a28df6636815fea0f55070999df9561fa9fa6