Add back an RST version of the project description
It seems pypi.python.org can't render markdown in the description,
which is probably why I had the RST version of the README (removed
in 007713d).

Add back an RST version, stripped down to minimal information about
the project. I don't want to have to maintain all the details of the
README in two places.

Change-Id: Ieb0a6c9cfe3f9cd9c976620b785ebbfdea1944e6