Allow data to be sent as `dict` and automatically convert it to `json`
There is confusion about what type of data should be passed with the
`data` parameter vs. with the `json` parameter (see #18).

The underyling requests library expects `json` to be a dict, and `data`
to be other types (usually a string). If a dict is passed to `data` it
results in a 400 Bad Request response from Gerrit.

Allow a dict to be passed to `data`, but before making the request,
replace it with `json`.

While writing tests for this functionality, the issue described in #29
was found. Stop using the _merge_dict method and explicitly reset the
`headers` dict before sending.

Fixes #18
Fixes #29

Change-Id: I27f1c0048535aa456c0a2cc6ace4cc893039ff62