Fixed an issue with duplicate output log message.
logging.BasicConfig() creating a StreamHandler with a default Formatter
and adding it to the root logger. Developers may unknowingly make the
logger contain duplicate handlers, leading to unexpected results such as
messages being duplicated in the log.

To maintain the log format, it is better to explicitly add streamhandler
to the named logger.

Signed-off-by: z-w-k <welcome.albert.z@gmail.com>