handles case where connection is  not perfect.
Sometimes the network is not as perfect as expected.
In that case, pygerrit2 behaviour is weird.
Here's the proposal is to set some retries (5) with a backoff_factor
(https://urllib3.readthedocs.io/en/latest/reference/urllib3.util.html#module-urllib3.util.retry)
and a timeout of 10s (gerrit should answer faster or not...)
with the retries and back off, requests will wait 0, 0.6, 1.2, 2.4 and
then 4.8s between the retries.

Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>