Align black and importlib_metadata python_version
Make the black [1] and importlib_metadata [2] python_version-s match in
Pipfile. Fix the local warning [3] with 'make testenvsetup' while doing
so (for importlib_metadata).

Now that importlib_metadata's python_version is marked as >3.5, remove
the inconsistent 3.5 from the Python versions listed in setup. This
removal being consistent also with the past commit b19e43c. Add 3.8 to
setup in turn, as that version should be listed so since commit 0805f99;
that also making this change fully consistent.

Add typing_extensions to packages, to fix CI's python 3.7 build of this
change. This change should end the recent series of CI build failures.

[1] https://pypi.org/project/black/, currently 21.5b1
[2] https://pypi.org/project/importlib-metadata/, currently 4.0.1
[3] Ignoring importlib-metadata: markers 'python_version < "3.8.0"' don't match your environment