Add 'six' to Pipfile dev-packages to fix livetests
Without this change, the livetests used to fail locally from within
python 3.9.5's docker package import (which was missing 'six'; [1]).

The failure is reproducible by removing the local virtualenv so that
'six' is nowhere to be found. Applying this change brings it in. This
change includes a Pipfile.lock that is regenerated from scratch, to
refresh it all while adding 'six' in.

[1] ModuleNotFoundError: No module named 'six'